+++ src/server/Model/Api/ScheduleModel.ts
@@ -13,7 +13,6 @@ interface ScheduleModelInterface extends ApiModel {
     getSchedule(time: number, length: number, type: apid.ChannelType): Promise<{}[]>;
     getScheduleId(time: number, channelId: number): Promise<{}>;
     getBroadcasting(addition: number): Promise<{}>;
-    getIPTVepg(days: number): Promise<string>;
     searchProgram(searchOption: SearchInterface): Promise<{}[]>;
     updateReserves(): Promise<void>;
 }
@@ -24,18 +23,18 @@ namespace ScheduleModelInterface {
 }
 
 class ScheduleModel extends ApiModel implements ScheduleModelInterface {
-    private programDB: ProgramsDBInterface;
-    private serviceDB: ServicesDBInterface;
+    private programsDB: ProgramsDBInterface;
+    private servicesDB: ServicesDBInterface;
     private ipc: IPCClientInterface;
 
     constructor(
-        programDB: ProgramsDBInterface,
-        serviceDB: ServicesDBInterface,
+        programsDB: ProgramsDBInterface,
+        servicesDB: ServicesDBInterface,
         ipc: IPCClientInterface,
     ) {
         super();
-        this.programDB = programDB;
-        this.serviceDB = serviceDB;
+        this.programsDB = programsDB;
+        this.servicesDB = servicesDB;
         this.ipc = ipc;
     }
 
@@ -48,8 +47,8 @@ class ScheduleModel extends ApiModel implements ScheduleModelInterface {
     */
     public async getSchedule(time: number, length: number, type: apid.ChannelType): Promise<{}[]> {
         let times = this.getTime(time, length);
-        let programs = await this.programDB.findSchedule(times.startAt, times.endAt, type);
-        let channels = await this.serviceDB.findChannelType([type]);
+        let programs = await this.programsDB.findSchedule(times.startAt, times.endAt, type);
+        let channels = await this.servicesDB.findChannelType([type]);
 
         // sort
         channels = ApiUtil.sortItems(channels, this.config.getConfig().serviceOrder || []);
@@ -89,10 +88,10 @@ class ScheduleModel extends ApiModel implements ScheduleModelInterface {
         let programs: DBSchema.ScheduleProgramItem[][] = [];
         for(let i = 0; i < 7; i++) {
             let addTime = i * 24 * 60 * 60 * 1000;
-            programs.push(await this.programDB.findScheduleId(times.startAt + addTime, times.endAt + addTime, channelId));
+            programs.push(await this.programsDB.findScheduleId(times.startAt + addTime, times.endAt + addTime, channelId));
         }
 
-        let channel = await this.serviceDB.findId(channelId);
+        let channel = await this.servicesDB.findId(channelId);
 
         if(channel === null) { throw new Error(ScheduleModelInterface.channelIdIsNotFoundError); }
 
@@ -110,8 +109,8 @@ class ScheduleModel extends ApiModel implements ScheduleModelInterface {
     * @return Promise<{}>
     */
     public async getBroadcasting(addition: number): Promise<{}> {
-        let programs = await this.programDB.findBroadcasting(addition * 1000 * 60);
-        let channels = await this.serviceDB.findAll();
+        let programs = await this.programsDB.findBroadcasting(addition * 1000 * 60);
+        let channels = await this.servicesDB.findAll();
 
         // sort
         channels = ApiUtil.sortItems(channels, this.config.getConfig().serviceOrder || []);
@@ -181,73 +180,6 @@ class ScheduleModel extends ApiModel implements ScheduleModelInterface {
     }
 
     /**
-    * Kodi IPTV 番組情報を生成
-    * @return Promise<string>
-    */
-    public async getIPTVepg(days: number): Promise<string> {
-        const now = new Date().getTime();
-        const programs = await this.programDB.findSchedule(now, now + 1000 * 60 * 60 * 24 * days);
-        const channels = await this.serviceDB.findAll();
-
-        // channelId ごとに programs をまとめる
-        let programsIndex: { [key: number]: DBSchema.ScheduleProgramItem[] } = {};
-        for(let program of programs) {
-            if(typeof programsIndex[program.channelId] === 'undefined') {
-                programsIndex[program.channelId] = [];
-            }
-
-            program.name = this.replaceStr(program.name);
-            if(program.description !== null) {
-                program.description = this.replaceStr(program.description);
-                if(program.extended !== null) {
-                    program.description += this.replaceStr(program.extended);
-                }
-            }
-            programsIndex[program.channelId].push(program);
-        }
-
-
-        let str = '<?xml version="1.0" encoding="UTF-8"?>'
-            + '<!DOCTYPE tv SYSTEM "xmltv.dtd">'
-            + '<tv generator-info-name="EPGStation">'
-        for(let channel of channels) {
-            if(typeof programsIndex[channel.id] === 'undefined') { continue; }
-            str += `<channel id="${ channel.id }" tp="${ channel.channel }">`;
-            str += `<display-name lang="ja_JP">${ channel.name }</display-name>`;
-            str += `<service_id>${  channel.serviceId }</service_id>`;
-            str += '</channel>\n';
-
-            for(let program of programsIndex[channel.id]) {
-                str += `<programme start="${ this.getTimeStr(program.startAt) }" stop="${ this.getTimeStr(program.endAt) }" channel="${ program.channelId }">`;
-                str += `<title lang="ja_JP">${ program.name }</title>`;
-                if(program.description !== null) { str += `    <desc lang="ja_JP">${ program.description }</desc>`; }
-                str += '</programme>';
-            }
-        }
-        str += '</tv>';
-
-        return str;
-    }
-
-    /**
-    * xml での禁止文字列を置き換える
-    * @param str: string
-    * @return string
-    */
-    private replaceStr(str: string): string {
-        return str.replace(/</g,'＜').replace(/>/g,'＞').replace(/&/g, '＆').replace(/"/g, '”').replace(/'/g, '’');
-    }
-
-    /**
-    * iptv 用の時刻文字列を生成する
-    * @param time: apid.UnixtimeMS
-    @return string
-    */
-    private getTimeStr(time: apid.UnixtimeMS): string {
-        return DateUtil.format(new Date(time), 'yyyyMMddhhmmss +0900');
-    }
-
-    /**
     * rule 検索
     * @param searchOption: SearchInterface
     * @return Promise<{}[]>
@@ -258,7 +190,7 @@ class ScheduleModel extends ApiModel implements ScheduleModelInterface {
             throw new Error(ScheduleModelInterface.searchOptionIsIncorrect);
         }
 
-        let programs = await this.programDB.findRule(searchOption, true, 300);
+        let programs = await this.programsDB.findRule(searchOption, true, 300);
 
         return programs.map((program) => {
             return ApiUtil.deleteNullinHash(program);
